+++ config/pipeline.php
@@ -25,15 +25,17 @@ use Psr\Container\ContainerInterface;
 
 return function (Application $app, MiddlewareFactory $factory, ContainerInterface $container): void {
     $app->pipe(Middleware\MetaMiddleware::class);
+    $app->pipe(Middleware\CorsHeaderMiddleware::class);
     $app->pipe(ErrorHandler::class);
     $app->pipe(Middleware\CleanupMiddleware::class);
 
     $app->pipe(BaseUrlMiddleware::class);
     $app->pipe(ServerUrlMiddleware::class);
     $app->pipe(RouteMiddleware::class);
-    $app->pipe(MethodNotAllowedMiddleware::class);
     $app->pipe(ImplicitHeadMiddleware::class);
     $app->pipe(ImplicitOptionsMiddleware::class);
+    $app->pipe(MethodNotAllowedMiddleware::class);
+
     $app->pipe(Middleware\AuthorizationMiddleware::class);
     $app->pipe(Middleware\RequestDeserializerMiddleware::class);
     $app->pipe(Middleware\ResponseSerializerMiddleware::class);