+++ config/autoload/base.global.php
@@ -17,5 +17,5 @@ return [
     ConfigAggregator::ENABLE_CACHE => true,
     'debug' => false,
     'name' => 'Factorio Item Browser - API Server',
-    'version' => '2.1.0',
+    'version' => '2.1.1',
 ];