+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Allow "base" mod to be absent in `/auth` request as of FFF #343.
+- Indirect dependency `dasprid/container-interop-doctrine` to `roave/psr-container-doctrine`.
 
 ## 2.0.0 - 2020-04-16